LayoutTests/fast/forms/select/popup-closes-on-blur-expected.txt

Test for http://bugs.webkit.org/show_bug.cgi?id=113220: ensures select popup closes when focus changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL internals.isSelectPopupVisible(popup) should be true. Was false.
PASS internals.isSelectPopupVisible(popup) is false
PASS successfullyParsed is true

TEST COMPLETE
